# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finlytic — financial document analysis & forecasting
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface KeyMetric {
  label: string;
  value: number;
  /** @nullable */
  unit?: string | null;
  /**
   * Period-over-period change as a fraction (e.g. 0.12 = +12%)
   * @nullable
   */
  change?: number | null;
}

export interface SeriesPoint {
  /** Time bucket label (e.g. "Q1 2024", "Jan", "2023") */
  period: string;
  value: number;
}

export interface TimeSeries {
  label: string;
  /** @nullable */
  unit?: string | null;
  points: SeriesPoint[];
}

export interface ForecastPoint {
  period: string;
  value: number;
  lower: number;
  upper: number;
}

export interface Forecast {
  label: string;
  /** Forecasting method used (e.g. "linear-regression") */
  method: string;
  history: SeriesPoint[];
  projection: ForecastPoint[];
}

export interface AnalysisSummary {
  id: number;
  filename: string;
  fileType: string;
  fileSize: number;
  title: string;
  summary: string;
  /** One of "positive", "neutral", "negative" */
  sentiment: string;
  createdAt: string;
}

export interface Analysis {
  id: number;
  filename: string;
  fileType: string;
  fileSize: number;
  title: string;
  summary: string;
  sentiment: string;
  keyMetrics: KeyMetric[];
  timeSeries: TimeSeries[];
  forecasts: Forecast[];
  insights: string[];
  rawTextPreview: string;
  createdAt: string;
}

export type AnalysesStatsSentimentBreakdown = {
  positive: number;
  neutral: number;
  negative: number;
};

export interface AnalysesStats {
  totalAnalyses: number;
  totalMetricsExtracted: number;
  totalForecasts: number;
  sentimentBreakdown: AnalysesStatsSentimentBreakdown;
  recent: AnalysisSummary[];
}

export type UploadAnalysisBody = {
  file: Blob;
  /** Number of forecast periods to generate (default 6) */
  horizon?: number;
};