# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Finlytic — financial document analysis & forecasting
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: analyses
    description: Financial document analysis operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /analyses:
    get:
      operationId: listAnalyses
      tags: [analyses]
      summary: List analyses
      description: Returns all analyses ordered by creation time, newest first.
      responses:
        "200":
          description: List of analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnalysisSummary"

  /analyses/upload:
    post:
      operationId: uploadAnalysis
      tags: [analyses]
      summary: Upload a financial document and run analysis
      description: |
        Accepts a PDF, CSV, XLSX or TXT file, parses the contents, runs an
        AI extraction pass to pull key metrics and time series, generates a
        forecast for any time series found, and persists the result.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                horizon:
                  type: integer
                  description: Number of forecast periods to generate (default 6)
      responses:
        "201":
          description: Analysis created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /analyses/{id}:
    get:
      operationId: getAnalysis
      tags: [analyses]
      summary: Get an analysis by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAnalysis
      tags: [analyses]
      summary: Delete an analysis
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /analyses/stats:
    get:
      operationId: getAnalysesStats
      tags: [analyses]
      summary: Aggregated stats across all analyses
      description: Lightweight roll-up used to populate dashboard headers.
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysesStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    KeyMetric:
      type: object
      required: [label, value]
      properties:
        label:
          type: string
        value:
          type: number
        unit:
          type: ["string", "null"]
        change:
          type: ["number", "null"]
          description: Period-over-period change as a fraction (e.g. 0.12 = +12%)

    SeriesPoint:
      type: object
      required: [period, value]
      properties:
        period:
          type: string
          description: Time bucket label (e.g. "Q1 2024", "Jan", "2023")
        value:
          type: number

    TimeSeries:
      type: object
      required: [label, points]
      properties:
        label:
          type: string
        unit:
          type: ["string", "null"]
        points:
          type: array
          items:
            $ref: "#/components/schemas/SeriesPoint"

    ForecastPoint:
      type: object
      required: [period, value, lower, upper]
      properties:
        period:
          type: string
        value:
          type: number
        lower:
          type: number
        upper:
          type: number

    Forecast:
      type: object
      required: [label, history, projection, method]
      properties:
        label:
          type: string
        method:
          type: string
          description: Forecasting method used (e.g. "linear-regression")
        history:
          type: array
          items:
            $ref: "#/components/schemas/SeriesPoint"
        projection:
          type: array
          items:
            $ref: "#/components/schemas/ForecastPoint"

    AnalysisSummary:
      type: object
      required:
        - id
        - filename
        - fileType
        - fileSize
        - title
        - summary
        - sentiment
        - createdAt
      properties:
        id:
          type: integer
        filename:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        title:
          type: string
        summary:
          type: string
        sentiment:
          type: string
          description: One of "positive", "neutral", "negative"
        createdAt:
          type: string
          format: date-time

    Analysis:
      type: object
      required:
        - id
        - filename
        - fileType
        - fileSize
        - title
        - summary
        - sentiment
        - keyMetrics
        - timeSeries
        - forecasts
        - insights
        - rawTextPreview
        - createdAt
      properties:
        id:
          type: integer
        filename:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        title:
          type: string
        summary:
          type: string
        sentiment:
          type: string
        keyMetrics:
          type: array
          items:
            $ref: "#/components/schemas/KeyMetric"
        timeSeries:
          type: array
          items:
            $ref: "#/components/schemas/TimeSeries"
        forecasts:
          type: array
          items:
            $ref: "#/components/schemas/Forecast"
        insights:
          type: array
          items:
            type: string
        rawTextPreview:
          type: string
        createdAt:
          type: string
          format: date-time

    AnalysesStats:
      type: object
      required:
        - totalAnalyses
        - totalMetricsExtracted
        - totalForecasts
        - sentimentBreakdown
        - recent
      properties:
        totalAnalyses:
          type: integer
        totalMetricsExtracted:
          type: integer
        totalForecasts:
          type: integer
        sentimentBreakdown:
          type: object
          required: [positive, neutral, negative]
          properties:
            positive:
              type: integer
            neutral:
              type: integer
            negative:
              type: integer
        recent:
          type: array
          items:
            $ref: "#/components/schemas/AnalysisSummary"