# lib/api-zod/src/generated/types/analysesStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finlytic — financial document analysis & forecasting
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysesStatsSentimentBreakdown } from "./analysesStatsSentimentBreakdown";
import type { AnalysisSummary } from "./analysisSummary";

export interface AnalysesStats {
  totalAnalyses: number;
  totalMetricsExtracted: number;
  totalForecasts: number;
  sentimentBreakdown: AnalysesStatsSentimentBreakdown;
  recent: AnalysisSummary[];
}