# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finlytic — financial document analysis & forecasting
 * OpenAPI spec version: 0.1.0
 */
import type { Forecast } from "./forecast";
import type { KeyMetric } from "./keyMetric";
import type { TimeSeries } from "./timeSeries";

export interface Analysis {
  id: number;
  filename: string;
  fileType: string;
  fileSize: number;
  title: string;
  summary: string;
  sentiment: string;
  keyMetrics: KeyMetric[];
  timeSeries: TimeSeries[];
  forecasts: Forecast[];
  insights: string[];
  rawTextPreview: string;
  createdAt: Date;
}