# lib/api-zod/src/generated/types/analysisSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finlytic — financial document analysis & forecasting
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisSummary {
  id: number;
  filename: string;
  fileType: string;
  fileSize: number;
  title: string;
  summary: string;
  /** One of "positive", "neutral", "negative" */
  sentiment: string;
  createdAt: Date;
}