# lib/api-zod/src/generated/types/forecast.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finlytic — financial document analysis & forecasting
 * OpenAPI spec version: 0.1.0
 */
import type { ForecastPoint } from "./forecastPoint";
import type { SeriesPoint } from "./seriesPoint";

export interface Forecast {
  label: string;
  /** Forecasting method used (e.g. "linear-regression") */
  method: string;
  history: SeriesPoint[];
  projection: ForecastPoint[];
}