# lib/api-zod/src/generated/types/keyMetric.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finlytic — financial document analysis & forecasting
 * OpenAPI spec version: 0.1.0
 */

export interface KeyMetric {
  label: string;
  value: number;
  /** @nullable */
  unit?: string | null;
  /**
   * Period-over-period change as a fraction (e.g. 0.12 = +12%)
   * @nullable
   */
  change?: number | null;
}