# lib/api-zod/src/generated/types/seriesPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finlytic — financial document analysis & forecasting
 * OpenAPI spec version: 0.1.0
 */

export interface SeriesPoint {
  /** Time bucket label (e.g. "Q1 2024", "Jan", "2023") */
  period: string;
  value: number;
}