# lib/api-zod/src/generated/types/timeSeries.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finlytic — financial document analysis & forecasting
 * OpenAPI spec version: 0.1.0
 */
import type { SeriesPoint } from "./seriesPoint";

export interface TimeSeries {
  label: string;
  /** @nullable */
  unit?: string | null;
  points: SeriesPoint[];
}